+++ R/read_cross2.R
@@ -40,7 +40,8 @@ read_cross2 <-
 function(file, quiet=TRUE)
 {
     if(length(grep("\\.zip$", file)) > 0) { # zip file
-        dir <- tempdir()
+        dir <- qtl2_temp_dir()
+
         if(is_web_file(file)) {
             tmpfile <- tempfile()
             if(!quiet) message(" - downloading ", file, "\n       to ", tmpfile)
@@ -53,20 +54,31 @@ function(file, quiet=TRUE)
         file <- path.expand(file)
         stop_if_no_file(file)
         unzipped_files <- utils::unzip(file, exdir=dir)
+
+        # remove temporary directory on exit
+        on.exit({
+            if(!quiet) message(" - cleaning up")
+            unlink(dir, recursive=TRUE)
+        }, add=TRUE)
+
+        # ignore "__MACOSX/._" files
+        unzipped_files <- grep("__MACOSX/._", unzipped_files, fixed=TRUE, invert=TRUE, value=TRUE)
+
         if(any(grepl("\\.yaml$", unzipped_files))) {
-            file <- unzipped_files[grep("\\.yaml$", unzipped_files)]
+            file <- grep("\\.yaml$", unzipped_files, value=TRUE)
+            if(length(file) > 1)
+                stop("The zip file contains multiple yaml files")
+            if(any(grepl("\\.json$", unzipped_files)))
+                warning("The zip file contains both YAML and JSON files; using the YAML file.")
         }
         else if(any(grepl("\\.json$", unzipped_files))) {
-            file <- unzipped_files[grep("\\.json$", unzipped_files)]
+            file <- grep("\\.json$", unzipped_files, value=TRUE)
+            if(length(file) > 1)
+                stop("The zip file contains multiple json files")
         }
         else {
             stop('No ".yaml" or ".json" control file found')
         }
-
-        on.exit({ # clean up when done
-            if(!quiet) message(" - cleaning up")
-            unlink(unzipped_files)
-        }, add=TRUE)
     }
 
     # directory containing the data
@@ -724,4 +736,13 @@ reorder_map_table <-
     pos <- suppressWarnings( as.numeric(map_tab[,pos_col]) )
 
     map_tab[ order(chr, pos, seq_len(nrow(map_tab))), , drop=FALSE]
+}
+
+# create unique temporary directory name
+qtl2_temp_dir <-
+    function(n_letters=15, initial_bit="qtl2")
+{
+    file.path(tempdir(),
+              paste0(initial_bit, "_", paste(sample(c(letters, 0:9), n_letters, replace=TRUE), collapse="")) )
+
 }